+++ source/main/index.js
@@ -1,5 +1,6 @@
 // @flow
 import os from 'os';
+import path from 'path';
 import { app, BrowserWindow, shell } from 'electron';
 import { client } from 'electron-connect';
 import { Logger } from './utils/logging';
@@ -10,7 +11,13 @@ import { createMainWindow } from './windows/main';
 import { installChromeExtensions } from './utils/installChromeExtensions';
 import { environment } from './environment';
 import mainErrorHandler from './utils/mainErrorHandler';
-import { launcherConfig, frontendOnlyMode } from './config';
+import {
+  launcherConfig,
+  frontendOnlyMode,
+  pubLogsFolderPath,
+  APP_NAME,
+  stateDirectoryPath,
+} from './config';
 import { setupCardano } from './cardano/setup';
 import { CardanoNode } from './cardano/CardanoNode';
 import { safeExitWithCode } from './utils/safeExitWithCode';
@@ -20,8 +27,10 @@ import { detectSystemLocale } from './utils/detectSystemLocale';
 import { ensureXDGDataIsSet } from './cardano/config';
 import { rebuildApplicationMenu } from './ipc/rebuild-application-menu';
 import { detectSystemLocaleChannel } from './ipc/detect-system-locale';
+import { getStateDirectoryPathChannel } from './ipc/getStateDirectoryPathChannel';
 import { CardanoNodeStates } from '../common/types/cardano-node.types';
 import type { CheckDiskSpaceResponse } from '../common/types/no-disk-space.types';
+import { logUsedVersion } from './utils/logUsedVersion';
 
 /* eslint-disable consistent-return */
 
@@ -64,6 +73,10 @@ const safeExit = async () => {
 
 const onAppReady = async () => {
   setupLogging();
+  logUsedVersion(
+    environment.version,
+    path.join(pubLogsFolderPath, `${APP_NAME}-versions.json`)
+  );
 
   const cpu = os.cpus();
   const platformVersion = os.release();
@@ -142,6 +155,10 @@ const onAppReady = async () => {
   detectSystemLocaleChannel.onRequest(() => Promise.resolve(systemLocale));
 
   getNumberOfEpochsConsolidated();
+
+  getStateDirectoryPathChannel.onRequest(() =>
+    Promise.resolve(stateDirectoryPath)
+  );
 
   mainWindow.on('close', async event => {
     Logger.info(