+++ storybook/stories/Staking.stories.js
@@ -1,7 +1,7 @@
 // @flow
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import { withKnobs, date } from '@storybook/addon-knobs';
+import { withKnobs, date, number } from '@storybook/addon-knobs';
 import { linkTo } from '@storybook/addon-links';
 import StoryLayout from './support/StoryLayout';
 import StoryProvider from './support/StoryProvider';
@@ -13,13 +13,16 @@ import StakingWithNavigation from '../../source/renderer/app/components/staking/
 import StakingDelegationCountdown from '../../source/renderer/app/components/staking/delegation-countdown/StakingDelegationCountdown';
 import StakingDelegationCenter from '../../source/renderer/app/components/staking/delegation-center/StakingDelegationCenter';
 import StakingStakePools from '../../source/renderer/app/components/staking/stake-pools/StakingStakePools';
-import StakingRewards from '../../source/renderer/app/components/staking/rewards/StakingRewards';
 import StakingEpochs from '../../source/renderer/app/components/staking/epochs/StakingEpochs';
 import StakingInfo from '../../source/renderer/app/components/staking/info/StakingInfo';
 
-const defaultStartDateTime = new Date('Jun 01 2019');
+import StakingRewards from './Staking-Rewards.stories';
+
+const defaultPercentage = 10;
+const defaultStartDateTime = new Date('2019-09-26');
 const startDateTimeKnob = (name, defaultValue) => {
   const stringTimestamp = date(name, defaultValue);
+
   return new Date(stringTimestamp).toISOString();
 };
 
@@ -34,25 +37,33 @@ const pageNames = {
 storiesOf('Staking', module)
   .addDecorator((story, context) => {
     const storyWithKnobs = withKnobs(story, context);
-
     const getItemFromContext = () => context.parameters.id;
+    let activeSidebarCategory = null;
+
+    if (context.parameters.id === 'countdown') {
+      activeSidebarCategory =
+        CATEGORIES_BY_NAME.STAKING_WITH_DELEGATION_COUNTDOWN.route;
+    } else {
+      activeSidebarCategory =
+        CATEGORIES_BY_NAME.STAKING_WITHOUT_DELEGATION_COUNTDOWN.route;
+    }
 
     return (
       <StoryDecorator>
         <StoryProvider>
           <StoryLayout
-            activeSidebarCategory={CATEGORIES_BY_NAME.STAKING.route}
+            activeSidebarCategory={activeSidebarCategory}
             storyName={context.story}
           >
-            {context.parameters.id !== 'countdown' ? (
+            {context.parameters.id === 'countdown' ? (
+              storyWithKnobs
+            ) : (
               <StakingWithNavigation
                 activeItem={getItemFromContext()}
                 onNavItemClick={linkTo('Staking', item => pageNames[item])}
               >
                 {storyWithKnobs}
               </StakingWithNavigation>
-            ) : (
-              storyWithKnobs
             )}
           </StoryLayout>
         </StoryProvider>
@@ -62,13 +73,14 @@ storiesOf('Staking', module)
   // ====== Stories ======
 
   .add(
-    'Start of decentralisation notification',
+    'Decentralization Start Info',
     () => (
       <div>
         <StakingDelegationCountdown
+          redirectToStakingInfo={linkTo('Staking', () => 'Info')}
           currentLocale="en-US"
           startDateTime={startDateTimeKnob(
-            'Delegation Start DateTime',
+            'Decentralization Start DateTime',
             defaultStartDateTime
           )}
         />
@@ -89,14 +101,25 @@ storiesOf('Staking', module)
     { id: 'stake-pools' }
   )
 
-  .add(pageNames.rewards, () => <StakingRewards name={pageNames.rewards} />, {
-    id: 'rewards',
-  })
+  .add(pageNames.rewards, StakingRewards, { id: 'rewards' })
 
   .add(pageNames.epochs, () => <StakingEpochs name={pageNames.epochs} />, {
     id: 'epochs',
   })
 
-  .add(pageNames.info, () => <StakingInfo name={pageNames.info} />, {
-    id: 'info',
-  });
+  .add(
+    pageNames.info,
+    () => (
+      <StakingInfo
+        percentage={number('Percentage', defaultPercentage, {
+          min: 0,
+          max: 100,
+          step: 1,
+          range: true,
+        })}
+      />
+    ),
+    {
+      id: 'info',
+    }
+  );